introduce GroupAboutContentWidget
